improve data pc data processing
Previously, the algorithm to post process data
would discard data if it detected an error and
keep reading. However, this was done in such a way
that future reads would be misaligned and it would
throw everything out until another error occured
which realigned it. This fixes that alignment
issue so we're able to capture a lot more of the
good data after an error.