initial implementation of python-only host pc code
This implementation creates a basic shell environment for user
configuration options and retrieves data and logs it. Plotting is not
yet implemented.

Restricting host PC code to python rather than python + C code (where
C code reads data from the device) simplifies the logic significantly
at the expense of slower reads. The slower reads may not be an issue
since the plotting is the major bottleneck anyway. However, if the
python-implemented reading does prove to be too slow it should be
possible to implement the reading in multithreaded
C (producer-consumer) and tie it into Python with Cython. In any
event, it makes more sense to call the C code from python rather than
the other way around. The previous implementation, which piped the
data to a Python subprocess from C, required an additional interface
to communicate configuration options from the C code to Python which
is more complicated than calling C functions from Python.