rewrite communication with radar in C
Interfacing with the radar in C is much faster than in Python. Rough
profiling gives about 3.5MB/s in C and 200kB/s in Python. The
performance critical part is parsing the sweep data from the raw
bytes.

The C functions are exposed to Python via Cython. Surprisingly, the C
code is the performance bottleneck, not the plotting. PyQtgraph is
able to plot at the full 3.5MB/s bandwidth.

As a result of the current performance, rewriting the processing on
the FPGA shouldn't have any noticeable improvement since the
bottleneck is parsing the raw bytes to sweeps. If that can be improved
the FPGA code may start to have some beneficial effect.

It's worth keeping in mind that we're currently about 20x off the max
possible performance given the USB 2.0 data rate of 60MB/s.