gateware: live reconfiguration, including output data supported
This is an initial implementation of a fully reconfigurable design. By
fully reconfigurable, I mean that the FPGA waits for configuration
data from the host PC and then is able to send any output data and
supports any configuration of the frequency synthesizer. However, the
design is not yet tested (and therefore likely contains bugs) and the
PC code to support it has not yet been written.