gateware top: fix simulation
This simulation is unfortunately less flexible than the previous
one. The FT2232H RX FIFO buffer acts like it doesn't have an output
register. This means I can't use my existing FIFO to simulate it. We
could create a custom FIFO for this, but that seems like overkill at
the moment.