gateware fir: fix enable and latency
The FIR module takes longer to flush all values from its internal
registers than its data output latency. When the first inputs are
registered, it expects most of its internal registers to be
untouched (left with a value of 0). This change keeps these registers
0 until enable is asserted, at which point the registers begin to
fill.

This also solves a related issue with the latency. The correct latency
is actual 1 clock enable cycle instead of 2. The previous use of 2 was
a false positive believe because of the above mentioned issue.